requirements.txt
django==2.2.24
django-bootstrap4==1.0.1
django-forms-bootstrap==3.1.0
django-countries==5.5
django_crispy_forms==1.8.1
django-select2==7.1.2
python-decouple==3.1
django-redis==4.11.0
django-countries-plus==1.3.1
django-smart-selects==1.5.4
django-pdb==0.6.2
Pillow==8.3.2
djangorestframework==3.12.2
django-filter==2.4.0
django-bootstrap-modal-forms==2.0.1
django-widget-tweaks==1.4.8
Markdown
django-gitlog==0.2
googletrans
openpyxl==2.4.11
django-excel-response==2.0.5
extcolors

